+++ modules/tls/src/snhttp/tls/params/ECC.scala
@@ -1,7 +1,7 @@
 package snhttp.tls.params
 
 /// TLS EC Point Formats
-enum ECCPointFormat(val code: Int, val name: String):
+enum ECCPointFormat(val code: Short, val name: String):
   case UNCOMPRESSED extends ECCPointFormat(0, "uncompressed")
   case AnsiX962CompressedPrime extends ECCPointFormat(1, "ansiX962_compressed_prime")
   case AnsiX962CompressedChar2 extends ECCPointFormat(2, "ansiX962_compressed_char2")
@@ -10,7 +10,7 @@ enum ECCPointFormat(val code: Int, val name: String):
 end ECCPointFormat
 
 /// TLS EC Curve Types
-enum ECCType(val code: Int, val name: String):
+enum ECCType(val code: Short, val name: String):
   case ExplicitPrime extends ECCType(1, "explicit_prime")
   case ExplicitChar2 extends ECCType(2, "explicit_char2")
   case NamedCurve extends ECCType(3, "named_curve")
